Use BigDecimal

+++ todo-app/src/main/java/com/sauraedron/factory/TaskFactory.java
@@ -21,7 +21,7 @@ public class TaskFactory {
 
         BeanUtils.copyProperties(taskRequest, task1);
         task1.setTaskId(UUID.randomUUID().toString());
-        task1.setOrder(TaskOrderTracker.next().intValue());
+        task1.setOrder(TaskOrderTracker.next());
         task1.setCreatedAt(OffsetDateTime.now());
         return task1;
     }
